feat(world): the zoomable-map spatial skeleton (phases 1 to 3)
Build the large-scale-to-fine map the owner asked for: a generated world
read at any zoom from a single whole-world glyph down to per-tile detail,
deterministic and observer-independent (design Parts 6, 14, 54).

- topology: widen Coord3 and FlatBounded axes i16 to i32 (Part 56,
  large-scale worlds), folding the cell key by FNV-1a now that a lossless
  64-bit pack no longer fits, and computing neighbours in i64 and squared
  distance in i128-saturate so an edge cell cannot overflow.
- lod: the chunk grid (ChunkCoord, CHUNK) and the one canonical
  level-of-detail quadtree over the tile map (Part 6.2), each node
  summarised by its dominant biome with a lowest-id tiebreak, the whole
  tree folded by StateHasher in canonical order so a seed yields a
  bit-identical tree on any machine.
- view: the Camera and the multi-zoom glyph reads (whole_map_frame and
  Camera::frame), pure functions of the tree that draw the world without
  writing canon (Principle 10), the GPU path a later swap of the same
  reads.
- an end-to-end integration test (tests/zoom.rs) and a runnable example
  (examples/zoom_map.rs) that walks the zoom ladder.

34 world tests pass, clippy -D warnings and rustdoc -D warnings clean,
the full workspace green.

+++ crates/world/src/worldgen.rs
@@ -84,7 +84,7 @@ impl TileMap {
         let mut tiles = Vec::with_capacity(w * h);
         for y in 0..topo.height {
             for x in 0..topo.width {
-                let (xi, yi) = (x as i32, y as i32);
+                let (xi, yi) = (x, y);
                 let elevation = fractal(
                     seed,
                     xi,
@@ -177,8 +177,8 @@ impl TileMap {
 
 /// Blend a latitude gradient (warm at the equator row, cold at the poles) with temperature
 /// noise, in `[0, ONE]`.
-fn blend_latitude(y: i32, height: i16, noise: Fixed, weight: Fixed) -> Fixed {
-    let h = height as i32;
+fn blend_latitude(y: i32, height: i32, noise: Fixed, weight: Fixed) -> Fixed {
+    let h = height;
     let mid = h / 2;
     let lat = if mid > 0 {
         let dist = (y - mid).abs();
@@ -228,8 +228,8 @@ mod tests {
     fn a_generated_world_has_a_mix_of_biomes() {
         let m = gen(0xEA27);
         let mut seen = std::collections::BTreeSet::new();
-        for y in 0..24i16 {
-            for x in 0..48i16 {
+        for y in 0..24i32 {
+            for x in 0..48i32 {
                 seen.insert(m.tile(Coord3::ground(x, y)).unwrap().biome);
             }
         }
